+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.3.0-dev (Apr 01, 2026)
 
+- Add configurable navigation menu options `rootCategoryId`, `maxDepth`([#1292](https://github.com/commerce-emu/storefront-next/pull/1292))
 - Fix Vite dev server warnings for font loading from public directory ([#1286](https://github.com/commerce-emu/storefront-next/pull/1286))
 - Restored Page Designer aspect definitions for PDP and PLP (@W-21888616)
 - Added fallback functionality for Core Region component and added two regions on AboutUs page (@W-21527572)